==> NER/01_data_mining.ipynb <==
import pandas as pd

with open("../data/conll003-englishversion/train/train.txt", 'r', encoding='utf-8') as file:
    data = file.read()

with open("../data/conll003-englishversion/test/test.txt", 'r', encoding='utf-8') as file:
    data = file.read()

def parse_conll(data_string):
    data = []
    sentence_id = 0
    
    lines = data_string.strip().split('\n')
    
    for line in lines:
        line = line.strip()
        
        if not line:
            sentence_id += 1
            continue
            
        if line.startswith("-DOCSTART-"):
            continue
            
        parts = line.split()
        
        if len(parts) >= 4:
            data.append({
                'sentence_id': sentence_id,
                'word': parts[0],
                'pos_tag': parts[1],   
                'chunk_tag': parts[2], 
                'ner_tag': parts[3]    
            })
            
    return pd.DataFrame(data)


df = parse_conll(data)
print(df.head(10))

df['ner_tag'].unique()

df.to_csv('../data/conll003-englishversion/train/train.csv')

df.to_csv('../data/conll003-englishversion/test/test.csv')

def format_data_for_llm(df: pd.DataFrame) -> list[dict]:
    sentences = []
    
    grouped = df.groupby('sentence_id')
    
    for _, group in grouped:
        words = list(group['word'])
        tags = list(group['ner_tag'])
        
        sentences.append({
            'words': words,
            'tags': tags
        })
        
    return sentences


llm_data = format_data_for_llm(df)
llm_data[0]

data_for_df = []

for i, sentence_dict in enumerate(llm_data):
    joined_sentence = " ".join(sentence_dict['words'])
    joined_tags = " ".join(sentence_dict['tags'])
    
    data_for_df.append({
        'sentence_id': i + 1,
        'sentence_text': joined_sentence,
        'ner_tags_str': joined_tags,
    })

df_sentences = pd.DataFrame(data_for_df)
df_sentences

df_sentences.to_csv('train_for_llm.csv')

df_sentences.to_csv('test_for_llm.csv')

df_test = pd.read_csv('../data/conll003-englishversion/test/test_for_llm.csv')
df_train = pd.read_csv('../data/conll003-englishversion/train/train_for_llm.csv')


df_train.info()
