==> NER/02_llm_labeling.ipynb <==
import pandas as pd
import sys
import os

module_path = os.path.abspath(os.path.join(os.getcwd(), os.pardir))
if module_path not in sys.path:
    sys.path.append(module_path)

from ai_service import label_dataset_ner, AiClient
ai_service = AiClient()

train = pd.read_csv('../data/conll003-englishversion/train/train_for_llm.csv')

Due to the high cost of labeling the entire dataset, I selected 40 proposals for a test run of all models, after which I will label the entire dataset with the best model.

sample_df = train.sample(n=min(40, len(train)), random_state=42).reset_index(drop=True)
sample_df.head(2)

## promt

tamplate = {
    "role": "system",
    "content": """
        You are an expert Named Entity Recognition (NER) model. 
        Your task is to extract entities with maximum precision and return only valid JSON. 
        **Never output anything except JSON.**
        
        Rules:
            - Labels allowed: PER (Person), ORG (Organization), LOC (Location), MISC (Miscellaneous).
            - If no entities found, return {{'entities': []}}
            
        JSON Structure Example:
        {{
          [
            {{
              entity: [Example Entity Name],\n"
              label: PER|ORG|LOC|MISC \n"
            }}
          ]
        }}
        Text:\n{text}
    """
}

## gpt-4o

res_df = label_dataset_ner(sample_df, tamplate, ai_service.apenai_call, 'gpt-4o')
res_df.to_csv("ner_train_gpt_4o.csv")

## gpt-4o-mini

res_df = label_dataset_ner(sample_df, tamplate, ai_service.apenai_call, model_name='gpt-4o-mini')
res_df.to_csv("ner_train_gpt_4o_mini.csv")

res_df.head()

## gemini-2.0-flash

res_df = label_dataset_ner(sample_df, tamplate, ai_service.gemini_call, model_name='gemini-2.0-flash')
res_df.to_csv("ner_train_gemini-2.0-flash.csv")

res_df.head()

## mistral-small-3.1-24b-instruct

data = pd.read_csv("ner_train_mistral-small-3.1-24b-instruct.csv")
data.head()

# HUMAN -> bio_to_entities 
(O O O O B-LOC O B-LOC O) -> [{'entity': '..', 'label': '...'}, ...

def bio_to_entities(sentence_text, ner_tags_str):
    tokens = sentence_text.split()
    tags = ner_tags_str.split()

    entities = []
    current_tokens = []
    current_class = None

    for tok, tag in zip(tokens, tags):
        if tag == "O": # end entity
            if current_tokens:
                entities.append({
                    "entity": " ".join(current_tokens),
                    "label": current_class
                })
                current_tokens = []
                current_class = None
            continue

        tag_type, label = tag.split("-", 1)

        if tag_type == "B":  # start new
            if current_tokens:
                entities.append({
                    "entity": " ".join(current_tokens),
                    "label": current_class
                })
            current_tokens = [tok]
            current_class = label

        elif tag_type == "I":  # add 
            if current_class == label: 
                current_tokens.append(tok)
            else:
                if current_tokens:
                    entities.append({
                        "entity": " ".join(current_tokens),
                        "label": current_class
                    })
                current_tokens = [tok]
                current_class = label

    if current_tokens:
        entities.append({
            "entity": " ".join(current_tokens),
            "label": current_class
        })

    return entities

sample_df['entities'] = sample_df.apply(
    lambda row: bio_to_entities(row['sentence_text'], row['ner_tags_str']),
    axis=1
)

sample_df.head()

sample_df = sample_df.drop(columns=['Unnamed: 0'])
sample_df.head()

sample_df.to_csv('ner_train_human.csv')

train = pd.read_csv('../data/conll003-englishversion/train/train_for_llm.csv')

train['entities'] = train.apply(
    lambda row: bio_to_entities(row['sentence_text'], row['ner_tags_str']),
    axis=1
)

display(train.head())

train.to_csv('../data/conll003-englishversion/train/train.csv')

train.info()
